feat: Add ES6 basic constants module
This commit adds a new module called `0-constants.js` to the ES6 basic project. The module exports three functions: `taskFirst()`, `getLast()`, and `taskNext()`. The `taskFirst()` function returns a constant string value. The use of `const` is preferred when possible. The `getLast()` function returns a string value. The `taskNext()` function combines a variable with the result of the `getLast()` function and returns the combined string value. The variable is declared using `let`. These changes are necessary to introduce constants and demonstrate the use of `const` and `let` in ES6.