Add ES6 basic arrow function module
The module exports a default function `getNeighborhoodsList()`. Inside the function, the `sanFranciscoNeighborhoods` array is initialized with two values: 'SOMA' and 'Union Square'. The function also defines an arrow function `addNeighborhood()` which takes a `newNeighborhood` parameter and adds it to the `sanFranciscoNeighborhoods` array. The updated array is then returned. These changes are necessary to demonstrate the use of arrow functions in ES6.